e3/requirements.txt
boto3==1.4.1
botocore==1.4.62
click==6.6
Jinja2==2.8
python-dateutil==2.5.3
requests==2.11.1
rrdtool==0.1.6
spur==0.3.19
